+++ aspectjx/src/main/groovy/com/hujiang/gradle/plugin/android/aspectjx/internal/concurrent/BatchTaskScheduler.groovy
@@ -19,14 +19,12 @@ import java.util.concurrent.Executors
 
 /**
  * class description here
- * @author simon
- * @version 1.0.0
- * @since 2018-04-04
+ * @author simon* @version 1.0.0* @since 2018-04-04
  */
 class BatchTaskScheduler {
 
     ExecutorService executorService
-    List< ? extends ITask> tasks = new ArrayList<>()
+    List<? extends ITask> tasks = new ArrayList<>()
 
     BatchTaskScheduler() {
         executorService = Executors.newScheduledThreadPool(Runtime.runtime.availableProcessors() + 1)
@@ -37,7 +35,14 @@ class BatchTaskScheduler {
     }
 
     void execute() {
-        executorService.invokeAll(tasks)
+        def all = executorService.invokeAll(tasks)
+        all.each {
+            def get = it.get()
+            if (get instanceof Throwable) {
+                // 发生异常，中断构建
+                throw get
+            }
+        }
 
         tasks.clear()
     }