+++ aspectjx/src/main/groovy/com/hujiang/gradle/plugin/android/aspectjx/internal/procedure/OnFinishedProcedure.groovy
@@ -31,8 +31,8 @@ class OnFinishedProcedure extends AbsProcedure {
 
     @Override
     boolean doWorkContinuously() {
-        println "~~~~~~~~~~~~~~~~~~~~onFinished"
-        ajxCache.commit()
+        project.logger.debug("~~~~~~~~~~~~~~~~~~~~onFinished")
+        variantCache.ajxCache.commit()
 
         return true
     }