+++ aspectjx/src/main/groovy/com/hujiang/gradle/plugin/android/aspectjx/internal/procedure/UpdateInputFilesProcedure.groovy
@@ -36,7 +36,7 @@ class UpdateInputFilesProcedure extends AbsProcedure {
 
     @Override
     boolean doWorkContinuously() {
-        println "~~~~~~~~~~~~~~~~~~~~update input files"
+        project.logger.debug("~~~~~~~~~~~~~~~~~~~~update input files")
         BatchTaskScheduler taskScheduler = new BatchTaskScheduler()
 
         transformInvocation.inputs.each { TransformInput input ->
@@ -45,7 +45,7 @@ class UpdateInputFilesProcedure extends AbsProcedure {
                     @Override
                     Object call() throws Exception {
                         dirInput.changedFiles.each { File file, Status status ->
-                            println "~~~~~~~~~~~~~~~~changed file::${status.name()}::${file.absolutePath}"
+                            project.logger.debug("~~~~~~~~~~~~~~~~changed file::${status.name()}::${file.absolutePath}")
 
                             variantCache.includeFileContentTypes = dirInput.contentTypes
                             variantCache.includeFileScopes = dirInput.scopes
@@ -83,15 +83,21 @@ class UpdateInputFilesProcedure extends AbsProcedure {
                         }
                         //如果include files 发生变化，则删除include输出jar
                         if (variantCache.incrementalStatus.isIncludeFileChanged) {
-                            File includeOutputJar = transformInvocation.outputProvider.getContentLocation("include", variantCache.contentTypes,
-                                    variantCache.scopes, Format.JAR)
+                            File includeOutputJar = transformInvocation.outputProvider.getContentLocation(
+                                    "include",
+                                    variantCache.contentTypes,
+                                    variantCache.scopes,
+                                    Format.JAR)
                             FileUtils.deleteQuietly(includeOutputJar)
                         }
 
                         //如果exclude files发生变化，则重新生成exclude jar到输出目录
                         if (variantCache.incrementalStatus.isExcludeFileChanged) {
-                            File excludeOutputJar = transformInvocation.outputProvider.getContentLocation("exclude", variantCache.contentTypes,
-                                    variantCache.scopes, Format.JAR)
+                            File excludeOutputJar = transformInvocation.outputProvider.getContentLocation(
+                                    "exclude",
+                                    variantCache.contentTypes,
+                                    variantCache.scopes,
+                                    Format.JAR)
                             FileUtils.deleteQuietly(excludeOutputJar)
                             AJXUtils.mergeJar(variantCache.excludeFileDir, excludeOutputJar)
                         }
@@ -106,10 +112,14 @@ class UpdateInputFilesProcedure extends AbsProcedure {
                     taskScheduler.addTask(new ITask() {
                         @Override
                         Object call() throws Exception {
-                            println "~~~~~~~changed file::${jarInput.status.name()}::${jarInput.file.absolutePath}"
+                            project.logger.debug("~~~~~~~changed file::${jarInput.status.name()}::${jarInput.file.absolutePath}")
 
                             String filePath = jarInput.file.absolutePath
-                            File outputJar = transformInvocation.outputProvider.getContentLocation(jarInput.name, jarInput.contentTypes, jarInput.scopes, Format.JAR)
+                            File outputJar = transformInvocation.outputProvider.getContentLocation(
+                                    jarInput.name,
+                                    jarInput.contentTypes,
+                                    jarInput.scopes,
+                                    Format.JAR)
 
                             if (jarInput.status == Status.REMOVED) {
                                 variantCache.removeIncludeJar(filePath)