+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### v2.0.5(2019-09-18)
+* Replace `variant.getJavaCompiler()` with `variant.getJavaCompileProvider()`
+
+### v2.0.5(2019-02-11)
+* Changed `println` usages for a SLF4J logger instance in DEBUG mode
+
+### v2.0.4(2018-10-16)
+* [FIX] issue[#120](https://github.com/HujiangTechnology/gradle_plugin_android_aspectjx/issues/120)
+* [FIX] issue[#118](https://github.com/HujiangTechnology/gradle_plugin_android_aspectjx/issues/118)
+
+### v2.0.2(2018-08-07)
+* [FIX]无法识别kotlin写的Aspect文件
+
 ### v2.0.1(2018-05-03)
 * 解决dexguard混淆兼容性问题(ClassNotFoundException)